extracted navbar into its own component
